Refactor pagination types from i32 to u64 across multiple providers and services
- Updated `page` and `total_items` fields in various structs to use u64 instead of i32 for better handling of larger values.
- Adjusted related functions and methods in providers such as Anilist, Audible, Google Books, TMDB, and others to accommodate the new types.
- Ensured consistency in pagination logic throughout the codebase, including calculations for `next_page`.

+++ crates/models/common/src/lib.rs
@@ -249,14 +249,14 @@ pub struct NamedObject {
 )]
 pub struct SearchInput {
     pub take: Option<u64>,
-    pub page: Option<i32>,
+    pub page: Option<u64>,
     pub query: Option<String>,
 }
 
 #[derive(PartialEq, Eq, Serialize, Deserialize, Debug, SimpleObject, Clone, Default)]
 pub struct SearchDetails {
-    pub total_items: i32,
-    pub next_page: Option<i32>,
+    pub total_items: u64,
+    pub next_page: Option<u64>,
 }
 
 #[derive(Debug, InputObject, Clone, Serialize, Deserialize)]
