+++ crates/providers/giant-bomb/src/models.rs
@@ -56,11 +56,11 @@ pub struct GiantBombResource {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct GiantBombSearchResponse<T> {
-    pub offset: i32,
+    pub offset: u64,
     pub error: String,
     pub results: Vec<T>,
-    pub number_of_page_results: i32,
-    pub number_of_total_results: i32,
+    pub number_of_page_results: u64,
+    pub number_of_total_results: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
