+++ crates/providers/manga-updates/src/lib.rs
@@ -129,7 +129,7 @@ struct PersonItemResponse {
 
 #[derive(Serialize, Deserialize, Debug)]
 struct MetadataSearchResponse<T> {
-    total_hits: i32,
+    total_hits: u64,
     results: Vec<T>,
 }
 
@@ -159,7 +159,7 @@ impl MangaUpdatesService {
 impl MediaProvider for MangaUpdatesService {
     async fn people_search(
         &self,
-        page: i32,
+        page: u64,
         query: &str,
         _display_nsfw: bool,
         _source_specifics: &Option<PersonSourceSpecifics>,
@@ -341,7 +341,7 @@ impl MediaProvider for MangaUpdatesService {
 
     async fn metadata_search(
         &self,
-        page: i32,
+        page: u64,
         query: &str,
         _display_nsfw: bool,
         _source_specifics: &Option<MetadataSearchSourceSpecifics>,
